Added rogue server tracking and VelagonesAPI
