+++ build.gradle.kts
@@ -8,6 +8,7 @@ plugins {
 }
 
 group = "com.runicrealms.velagones"
+
 version = "0.1.7"
 
 repositories { mavenCentral() }
