+++ paper/src/main/kotlin/com/runicrealms/velagones/paper/AgonesHook.kt
@@ -7,11 +7,6 @@ import java.util.concurrent.ExecutorService
 import java.util.concurrent.Executors
 import java.util.concurrent.ScheduledExecutorService
 import net.infumia.agones4j.Agones
-import org.bukkit.Bukkit
-import org.bukkit.event.EventHandler
-import org.bukkit.event.Listener
-import org.bukkit.event.player.PlayerJoinEvent
-import org.bukkit.event.player.PlayerQuitEvent
 import org.slf4j.Logger
 
 class AgonesHook @Inject constructor(private val logger: Logger) {
