+++ velocity/src/main/kotlin/com/runicrealms/velagones/velocity/AutoscalerEndpoint.kt
@@ -15,7 +15,7 @@ class AutoscalerEndpoint(
     private val logger: Logger,
     private val autoscaler: Autoscaler,
     private val autoscalerConfig: AutoscalerConfig,
-    private val registry: VelagonesFleet.Registry,
+    private val registry: VelagonesGroup.Registry,
 ) {
 
     private val port =
